Authentication, design changes

+++ Application/Controllers/PatientController.cs
@@ -18,14 +18,17 @@ namespace Application.Controllers
         [HttpGet]
         public IActionResult GetAllPatients()
         {
-            IEnumerable<PatientDTO> patients = _patientService.GetAllPatients();
+            IEnumerable<PatientDto> patients = _patientService.GetAllPatients();
             return Ok(patients);
         }
 
         [HttpGet("{patientId}")]
         public IActionResult GetPatientById(int patientId)
         {
-            PatientDTO patientDTO = _patientService.GetPatientById(patientId);
+            if (patientId < 0)
+                return BadRequest();
+
+            PatientDto patientDTO = _patientService.GetPatientById(patientId);
             if (patientDTO == null)
             {
                 return NotFound();
@@ -36,14 +39,20 @@ namespace Application.Controllers
         [HttpPost]
         public IActionResult CreatePatientRecord(PatientCreationDTO patientCreationDTO)
         {
+            if (patientCreationDTO == null || !ModelState.IsValid)
+                return BadRequest();
+
             var patientCreatedDTO = _patientService.CreatePatientRecord(patientCreationDTO);
             return CreatedAtAction(nameof(GetPatientById), new { patientId = patientCreatedDTO.Id }, patientCreatedDTO);
         }
 
 
         [HttpPut]
-        public IActionResult UpdatePatientRecord(PatientDTO patientDTO)
+        public IActionResult UpdatePatientRecord(PatientDto patientDTO)
         {
+            if (patientDTO == null || !ModelState.IsValid)
+                return BadRequest();
+
             var patientUpdatedDTO = _patientService.UpdatePatientRecord(patientDTO);
             if (!patientUpdatedDTO)
             {
@@ -55,6 +64,9 @@ namespace Application.Controllers
         [HttpDelete]
         public IActionResult DeletePatientRecord(int patientId)
         {
+            if (patientId < 0)
+                return BadRequest();
+
             bool patientDeleted = _patientService.Delete(patientId);
             if (!patientDeleted)
             {
